pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "onion-guardian"
version = "0.1.0"
description = "Pre-execution security middleware with semantic policy inference for AI-powered IDEs"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Onion Guardian Contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
keywords = ["security", "middleware", "sandbox", "llm", "guardian", "ai-ide"]
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
guardian-local = [
    "transformers>=4.40",
    "torch>=2.0",
]
guardian-api = [
    "openai>=1.0",
    "anthropic>=0.20",
]
dev = [
    "build>=1.2",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.hatch.build]
include = [
    "LICENSE",
    "README.md",
    "onion_guardian/**/*.py",
    "onion_guardian/py.typed",
    "onion_guardian/config/*.yaml",
]

[tool.hatch.build.targets.wheel]
packages = ["onion_guardian"]

[project.urls]
Homepage = "https://github.com/odincon/onion-guardian"
Repository = "https://github.com/odincon/onion-guardian"
Issues = "https://github.com/odincon/onion-guardian/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
